Extract logic in new helper classes
